Start splitting index out of driver
