+++ datacube/index/common.py
@@ -1,0 +1,2 @@
+def _readable_offset(offset):
+    return '.'.join(map(str, offset))
